fix warnings

+++ src/query.c
@@ -317,7 +317,7 @@ int autoCreate(char *c_area, hs_addr pktOrigAddr, ps_addr forwardAddr)
     }
     i = ftell(f); /* config length */
     /*  add line to config */
-    if (fprintf(f, "%s%s", buff, cfgEol()) != strlen(buff)+strlen(cfgEol()) ||
+    if (fprintf(f, "%s%s", buff, cfgEol()) != (int)(strlen(buff)+strlen(cfgEol())) ||
         fflush(f) != 0) {
 	w_log(LL_ERR, "Error creating area %s, config write failed: %s!",
 	      c_area, strerror(errno));
